+++ internal/metrics/noop.go
@@ -9,9 +9,9 @@ import (
 
 type noopClient struct{}
 
-func (c noopClient) Gauge(string, float64, ...Tag) {}
+func (c noopClient) Gauge(string, int64, ...Tag) {}
 
-func (c noopClient) GaugeAdd(string, float64, ...Tag) {}
+func (c noopClient) GaugeAdd(string, int64, ...Tag) {}
 
 func (c noopClient) Count(string, int64, ...Tag) {}
